Hide DefaultUrlResolver implementation, move util functions to impl. file
